# docs/openapi.yaml
...
openapi: "3.0.3"
info:
  title: FamilyHub API
  version: "1.0"
  description: >
    Private, GEDCOM-7-backed family portal. All family content requires a
    session cookie (login). The /api/* surface is the WP2 JSON CRUD + search
    contract Cowork builds the WP3 front-end against; reads need any logged-in
    member, writes need at least the USER role (§10). State-changing /api calls
    must send the CSRF token in an X-CSRFToken header. The auth/admin/plumbing
    routes render HTML today; v2 can serve JSON from the same URLs.
...
paths:
...
  /api/media:
    get:
      tags: [Media]
      summary: List media (optionally for one subject)
      parameters: [{name: subject_type, in: query, schema: {type: string, enum: [individual, family, event]}}, {name: subject_id, in: query, schema: {type: integer}}]
      responses: {"200": {description: OK, content: {application/json: {schema: {type: object, properties: {media: {type: array, items: {$ref: '#/components/schemas/Media'}}}}}}}, "401": {$ref: '#/components/responses/Unauthorized'}}
    post:
      tags: [Media]
      summary: Upload an image (multipart); EXIF/GPS stripped, stored outside web root
      requestBody: {content: {multipart/form-data: {schema: {type: object, required: [file], properties: {file: {type: string, format: binary}, title: {type: string}, description: {type: string}, subject_type: {type: string}, subject_id: {type: integer}}}}}}
      responses: {"201": {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Media'}}}}, "400": {$ref: '#/components/responses/BadRequest'}, "403": {$ref: '#/components/responses/Forbidden'}}
  /api/media/{media_id}:
    parameters: [{name: media_id, in: path, required: true, schema: {type: integer}}]
    get: {tags: [Media], summary: One media object (with links), responses: {"200": {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/Media'}}}}, "404": {$ref: '#/components/responses/NotFound'}}}
    put: {tags: [Media], summary: Update metadata (title/description/gedcom_xref; bytes are immutable), requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/Media'}}}}, responses: {"200": {description: Updated}}}
    delete: {tags: [Media], summary: Delete the media object + its files + links, responses: {"204": {description: Deleted}, "404": {$ref: '#/components/responses/NotFound'}}}
  /api/media/{media_id}/file:
    parameters: [{name: media_id, in: path, required: true, schema: {type: integer}}]
    get: {tags: [Media], summary: Full image bytes (login-walled), responses: {"200": {description: image bytes}, "401": {$ref: '#/components/responses/Unauthorized'}, "404": {$ref: '#/components/responses/NotFound'}}}
  /api/media/{media_id}/thumb:
    parameters: [{name: media_id, in: path, required: true, schema: {type: integer}}]
    get: {tags: [Media], summary: Thumbnail bytes (login-walled), responses: {"200": {description: image bytes}, "404": {$ref: '#/components/responses/NotFound'}}}
  /api/media/{media_id}/links:
    parameters: [{name: media_id, in: path, required: true, schema: {type: integer}}]
    post: {tags: [Media], summary: Attach the photo to a subject, requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/Link'}}}}, responses: {"201": {description: Linked}, "400": {$ref: '#/components/responses/BadRequest'}, "409": {description: Already linked}}}
  /api/media/{media_id}/links/{subject_type}/{subject_id}:
    parameters: [{name: media_id, in: path, required: true, schema: {type: integer}}, {name: subject_type, in: path, required: true, schema: {type: string}}, {name: subject_id, in: path, required: true, schema: {type: integer}}]
    delete: {tags: [Media], summary: Remove an attachment, responses: {"204": {description: Removed}, "404": {$ref: '#/components/responses/NotFound'}}}

  # --- Notes ------------------------------------------------------------------
...
components:
  responses:
    Unauthorized: {description: Not logged in, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Forbidden: {description: Logged in but lacking the required role, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    NotFound: {description: No such record, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    BadRequest: {description: Validation failed, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
    Error:
      type: object
      properties:
        error: {type: string}
        fields: {type: object, additionalProperties: {type: string}, description: per-field problems (present on validation errors)}
...
    Link:
      type: object
      description: A polymorphic attachment target.
      required: [subject_type, subject_id]
      properties:
        subject_type: {type: string, enum: [individual, family, event]}
        subject_id: {type: integer}
        subject_label: {type: string, nullable: true, readOnly: true}
    Media:
      type: object
      properties:
        id: {type: integer, readOnly: true}
        gedcom_xref: {type: string, nullable: true}
        file_path: {type: string, readOnly: true}
        media_type: {type: string, readOnly: true, example: image/jpeg}
        title: {type: string, nullable: true}
        description: {type: string, nullable: true}
        uploaded_by: {type: integer, nullable: true, readOnly: true}
        uploader: {type: string, nullable: true, readOnly: true}
        created_at: {type: string, format: date-time, readOnly: true}
        file_url: {type: string, readOnly: true}
        thumb_url: {type: string, readOnly: true}
        links_count: {type: integer, readOnly: true}
        links: {type: array, items: {$ref: '#/components/schemas/Link'}}
...